Move graph generation to a separate project
