+++ core/src/main/scala/com/mesosphere/usi/core/SchedulerFactory.scala
@@ -1,6 +1,7 @@
 package com.mesosphere.usi.core
 
 import akka.NotUsed
+import akka.stream.{FanInShape2, Graph}
 import akka.stream.scaladsl.Flow
 import com.mesosphere.mesos.client.MesosClient
 import com.mesosphere.usi.core.conf.SchedulerSettings
@@ -11,15 +12,16 @@ import com.mesosphere.usi.core.util.DurationConverters
 import com.mesosphere.usi.metrics.Metrics
 import com.mesosphere.usi.repository.PodRecordRepository
 import org.apache.mesos.v1.scheduler.Protos
-
+import org.apache.mesos.v1.scheduler.Protos.{Event => MesosEvent}
 import scala.concurrent.Future
 
 trait SchedulerLogicFactory {
-  def newSchedulerLogicGraph(snapshot: StateSnapshot): SchedulerLogicGraph
+  private[usi] def newSchedulerLogicGraph(
+      snapshot: StateSnapshot): Graph[FanInShape2[SchedulerCommand, MesosEvent, SchedulerEvents], NotUsed]
 }
 
 trait PersistenceFlowFactory {
-  def newPersistenceFlow(): Flow[SchedulerEvents, SchedulerEvents, NotUsed]
+  private[usi] def newPersistenceFlow(): Flow[SchedulerEvents, SchedulerEvents, NotUsed]
 }
 
 trait SuppressReviveFactory {
