+++ core/src/test/resources/application.conf
@@ -17,7 +17,3 @@ akka {
   loglevel = "DEBUG"
   logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
 }
-travesty {
-    cache.shape-cache-ttl = 5 minutes
-    engines = [CommandLine, JDK]
-}
