+++ docs/scheduler-vis/src/main/resources/application.conf
@@ -1,0 +1,4 @@
+travesty {
+    cache.shape-cache-ttl = 5 minutes
+    engines = [CommandLine, JDK]
+}
